Fixed signUp / FirebaseContext funcionality.
