+++ context/FirebaseContext.js
@@ -1,6 +1,6 @@
 import React, {useEffect, useState, useContext} from 'react';
 import { createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut, sendPasswordResetEmail, updateProfile} from 'firebase/auth';
-import { ref, update, get, child} from 'firebase/database';
+import { ref, update, get, set, child} from 'firebase/database';
 import { auth, database } from '../firebaseConfig';
 
 const FirebaseContext = React.createContext();
@@ -10,27 +10,36 @@ export function useFirebase() {
 }
 
 export function FirebaseProvider({children}) {
+    const [loading, setLoading] = useState(true);
     const [user, setUser] = useState(null); // Contains auth user object
     const [userData, setUserData] = useState(null); // Contains data for user in object format {fullName: string, rewardsPoints: int, hasSignUpReward: boolean, userEmail: string}
     const [hoursData, setHoursData] = useState(null);
 
     // Function to sign a user up. Calls firebase method to create auth user, and then uses this new user to update the database and local state data.
-    function signUp(email, password, name) {
-        return createUserWithEmailAndPassword(auth, email, password)
-            .then((userCredential) => {
-                const newUser = userCredential.user;
-                const userRef = ref(database, `users/${newUser.uid}`);
-                updateUserName(name)
-                .then(() => {
-                    update(userRef, {fullName: name, rewardsPoints: 0, hasSignUpReward: true, userEmail: email})
-                    .then(() => {
-                        setUser(newUser);
-                        setUserData({name: newUser.displayName, email: newUser.email, rewardsPoints: 0, hasSignUpReward: true})
-                        console.log(userData);
-                        return newUser;
-                    })
-                })
-            });
+    async function signUp(email, password, name) {
+        try {
+            setLoading(true)
+            // Create a new user with email and password
+            createUserWithEmailAndPassword(auth, email, password)
+                .then((userCredential) => {
+                    const newUser = userCredential.user;
+                    updateProfile(newUser, {displayName: name}).then(() => {
+                        const userRef = ref(database, `users/${newUser.uid}`)
+                        set(userRef, {
+                            fullName: name,
+                            rewardsPoints: 0,
+                            hasSignUpReward: true,
+                            userEmail: email,
+                        }).then(() => {
+                            setUserData({fullName: name, rewardsPoints: 0, hasSignUpReward: true, userEmail: email})
+                            return newUser;
+                        })
+                    });
+                });
+            
+        } catch (error) {
+            console.error('failed, '+  error);
+        }
     }
 
     // Logs a user into the platform. Calls firebase method to sign in, and then accesses the database to update local data. 
@@ -43,7 +52,6 @@ export function FirebaseProvider({children}) {
                 get(child(userDataRef, `users/${loggedUser.uid}`)).then((snapshot) => {
                     if (snapshot.exists()) {
                         setUserData(snapshot.val());
-                        console.log(snapshot.val());
                     } else {
                         console.log('No data available')
                     }
@@ -111,13 +119,14 @@ export function FirebaseProvider({children}) {
                 // User is signed out.
                 setUser(null);
             }
+            setLoading(false);
         })
 
         return unsubscribe;
     }, []);
 
     return (
-        <FirebaseContext.Provider value={{user, userData, hoursData, signUp, logIn, logOut, updateUserName, resetPassword, redeemReward, getUserData}}>
+        <FirebaseContext.Provider value={{loading, user, userData, hoursData, signUp, logIn, logOut, updateUserName, resetPassword, redeemReward, getUserData}}>
             {children}
         </FirebaseContext.Provider>
     )
